Resolved one more warning in dorado/polish/medaka_read_matrix.cpp.

+++ dorado/polish/medaka_read_matrix.cpp
@@ -290,7 +290,7 @@ ReadAlignmentData calculate_read_alignment(BamFile &bam_file,
     // allocate output assuming one insertion per ref position
     int32_t n_pos = 0;
     int32_t max_n_reads = 0;
-    const int32_t buffer_pos = 2 * (end - start);
+    const int32_t buffer_pos = static_cast<int32_t>(2 * (end - start));
     const int32_t buffer_reads = std::min(max_reads, 100);
     const int32_t extra_featlen =
             (include_dwells ? 1 : 0) + (include_haplotype ? 1 : 0) + (num_dtypes > 1 ? 1 : 0);
